% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnTranslation.R
\name{columnTranslation}
\alias{columnTranslation}
\title{Column Name Translation Function}
\usage{
columnTranslation(file, rawData)
}
\arguments{
\item{file}{A character string specifying the path to the file whose column names need to be translated.}

\item{rawData}{A data frame whose column names will be translated.}
}
\value{
A data frame with the translated column names.
}
\description{
This function translates column names in \code{rawData} based on the file type detected from the input file.
It uses the column name mappings from the JSON file.
}
