% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAATable.R
\name{getAATable}
\alias{getAATable}
\title{getAATable}
\usage{
getAATable(matrixType = "SER")
}
\arguments{
\item{matrixType}{Character for the type of matrix. Often serum ("SER"),
plasma ("PLA") or urine ("URI").}
}
\value{
A data table with information about amino acids.
\describe{
\item{\code{Compound}}{Name of the compound.}
\item{\code{Unit}}{Unit of measurement taken from instrument SOP.}
\item{\code{Max Value (ref.)}}{Biologically based maximum value for a compound. Based on regular population.}
\item{\code{Min Value (ref.)}}{Biologically based minimum value for a compound. Based on regular population.}
\item{\code{Reference Unit}}{Unit of measurement taken from instrument SOP.}
\item{\code{Reference Range}}{Reference minimum to reference maximum already in the table.}
}
}
\description{
Gets the names and description of Amino Acids for publication purposes.
}
