% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longFormat.R
\name{longFormat}
\alias{longFormat}
\title{longFormat}
\usage{
longFormat(rawData)
}
\arguments{
\item{rawData}{data in wide format}
}
\value{
data frame in long format. "sampleID", "AnalyteName", "AnalysisName",
"cohortName","projectName","sampleMatrixType", "sampleType" are still columns
but the rest are stored in paramName and paramValue.
}
\description{
Takes data that has been parsed and tidied and puts it into a long format for
database storage.
}
