% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms.parser-package.R
\docType{package}
\name{ms.parser-package}
\alias{ms.parser}
\alias{ms.parser-package}
\title{ms.parser: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
