% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAA.R
\name{readAA}
\alias{readAA}
\title{Read amino acid data from tsv's}
\usage{
readAA(file, optns = list())
}
\arguments{
\item{file}{\itemize{
\item the file path to be imported
}}

\item{optns}{list of options
\itemize{
\item codePosition - position of the code in the file name for sampleID.
Default is where cal is found in the analysis name.
\item projectName - the name of the project. For example "covid19". The default
will take it from the first position in AnalysisName.
\item cohortName - the name of the cohort. For example "harvardC2". The default
will take it from the second position in AnalysisName
\item sampleMatrixType - sample matrix type, usually "PLA", "SER" or "URI".
The default will take it from the AnalysisName
}}
}
\value{
rawData read from TSV file
}
\description{
Read amino acid data from tsv's
}
