% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSCFA.R
\name{readSCFA}
\alias{readSCFA}
\title{extract targeted MS data from task export xml file}
\usage{
readSCFA(path, optns = list())
}
\arguments{
\item{path}{\itemize{
\item the path to the expName folder
}}

\item{optns}{list of optns
\itemize{
\item columnList - list of columns to be selected
}}
}
\value{
a dataElement
}
\description{
extract targeted MS data from task export xml file
}
