% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTRY.r
\name{readTRY}
\alias{readTRY}
\title{extract targeted MS data from task export xml file}
\usage{
readTRY(path, optns = list())
}
\arguments{
\item{path}{\itemize{
\item the path to the expName folder
}}

\item{optns}{list of optns
\itemize{
\item codePosition - position of the code in the file name. Default is 8.
\item columnList - list of columns to be selected
}}
}
\value{
a dataElement
}
\description{
extract targeted MS data from task export xml file
}
