% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTXT.R
\name{readTXT}
\alias{readTXT}
\title{readTXT
Reads .txt outputs for mass spectrometry}
\usage{
readTXT(path)
}
\arguments{
\item{path}{character for path to file}
}
\value{
data frame with no empty or all zero columns
}
\description{
readTXT
Reads .txt outputs for mass spectrometry
}
