% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readXML2.R
\name{readXML2}
\alias{readXML2}
\title{readXML2}
\usage{
readXML2(path)
}
\arguments{
\item{path}{character for path to file}
}
\value{
data frame with no empty or all zero columns
}
\description{
Reads xml outputs for mass spectrometry.
}
