% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAA.R
\name{readAA}
\alias{readAA}
\title{Read amino acid data from tsv's}
\usage{
readAA(file, optns = list())
}
\arguments{
\item{file}{\itemize{
\item the file path to be imported
}}

\item{optns}{list of options
\itemize{
\item codePosition - position of the code in the file name. Default is
where cal is found in the analysis name.
}}
}
\value{
rawData read from TSV file
}
\description{
Read amino acid data from tsv's
}
